Revert "Clean up redis configuration. Allow using REDIS_URL to set advanced (#2732)"
REDIS_URL was intended to be a single exhaustive configuration for Redis,
but it was inappropriate because of incompatibility of Unix socket URLs.

This reverts commit c997091166ba1801eea3a587c913b020b9b84ce4.

The specification of redis scheme does not specify about transporting
protocol.
https://www.iana.org/assignments/uri-schemes/prov/redis

Both of redis-rb 3.2 and Node.js redis 2.6.5, which are dependents, assumes the transporting
protocol is TCP by default, but they provide different syntax to change
the configuration.

redis-rb 3.2 has "unix" scheme to configure Unix socket. The scheme behaves
same with redis scheme except the default transporation is Unix socket.
https://github.com/redis/redis-rb/blob/v3.2.2/lib/redis/client.rb#L393

Node.js redis 2.6.5, on the other hand, regards the URL as the path to a
Unix socket if the URL lacks the scheme and succeeding two slashes.
https://github.com/NodeRedis/node_redis/blob/v.2.6.5/lib/createClient.js#L29

Because of the difference, using URL is insufficient to add an
intended feature, Unix socket support. It is even problematic for
environments which have already used the former configuration variables and
a user-defined configuration variable for Unix socket support.

This commit reverts the change and maintains the old configuration until
we make another change considering the concern.