Introduce methods for bulk operations to FeedManager.
This commit substitutes filter? of FeedManager with filter_subscribers
and filter_mentions, which filter relations for ease of use and better
performance. It substitutes also push and trim with push_bulk and
trim_bulk, respectively.

It removes FeedInsertWorker, whose main role was to call filter?
for each insertions.

It removes filter? call from NotifyService.

It adds the following scopes to AccountInteractions: following,
not_blocking_accounts_in, not_blocked_by, not_domain_blocking, and
not_muting_accounts_in.

It removes PushUpdateWorker and now the feature is divided into
push_bulk and publish of FeedManager and PushReblogWorker. The
same function could be implemented by push_bulk, and so does
FanOutOnWriteService.

filter? was a method to determine a status to be filtered or not.
However, callers of filter? actually deal with relations and it is
favorable to filter relations rather than requiring to call a method
for each records in the relations. That could also benefit for
performance, by reducing the linear query number to a constant number.

The main role of FeedInsertWorker was to call filter? for a status, and
jobs for it was queued by FanOutOnWriteService for each statuses. Now
filter_subscribers filters multiple statuses rather than a status, so
remove FeedInsertWorker and move the feature to FanOutOnWriteService.

This commit removes filter? call from NotifyService. Now the callers filter
mentions, which are in relations, by themselves.

PushUpdateWorker was basically a worker to make a request to Redis for each
accounts. Now the requests are pipelined in push_bulk by Redis, which
achieves concurrency. That removes the overhead of Sidekiq.

On the other hand, it prevents from constructing pipelines for the whole
data stream. In the new implementation, an operation for each timeline has
to wait for operations for other timelines while it did not have to in the
old one.