Use nullDB during asset compilation
This change allows Nanobox users to continue deploying Mastodon instances, and also speeds up the asset compilation phase for other users who might want it. It adds the ActiveRecord nullDB driver, and lets users switch to it by setting an environment variable (`DB_ADAPTER=nulldriver`), which can be done as part of the `rake assets:precompile` command. For Nanobox users, it relieves the need for an actual database connection during the `compile` step (part of the deploy process which prepares the code for upload to the server). For all users, it speeds up asset compilation by converting all DB calls to no-ops, which is considerably faster than talking to an actual Postgres DB during a phase which doesn't actually need DB access to begin with (even though Rails loads itself entirely for the process anyway).